Refactor details panel (#8551)
* moved citation-box and util to new dir

* move point of contact to own component

* renamed details directory
