+++ src/pages/public_registration/details-panel/_components/citation-box/CitationBox.tsx
@@ -1,8 +1,8 @@
 import { TextField, Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
-import { Registration } from '../../../types/registration.types';
-import { useJournalSeoData } from '../../../utils/hooks/useJournalSeoData';
-import { formatAPA } from './format-apa';
+import { Registration } from '../../../../../types/registration.types';
+import { useJournalSeoData } from '../../../../../utils/hooks/useJournalSeoData';
+import { formatAPA } from './_utils/format-apa';
 
 const citationHeadingId = 'citation-box-heading';
